Implementing bcryptjs and JWT for password hashing in Node JS
This commit implements password hashing using bcryptjs and JSON Web Token (JWT) in Node JS. The bcryptjs library is used to securely hash passwords before storing them in the database. This implementation ensures that users' passwords are not stored in plaintext, providing an extra layer of security. Additionally, JWT is implemented to generate an authentication token upon successful login. This token is then used to authenticate subsequent requests from the user, improving the security of the application. Overall, this implementation enhances the security of the application's password storage and user authentication process.