Creating Login Endpoints and Sending Auth Token
This commit adds login endpoints to the project and sends an authentication token upon successful authentication. With these new endpoints, users can securely authenticate themselves to the application. Upon successful authentication, the server generates an authentication token using JSON Web Token (JWT) and sends it to the client. This token can be used to authenticate the user's subsequent requests to the server. This implementation enhances the security of the application's user authentication process, providing a seamless and secure experience for the users.