Adding middleware to decode JWT and get user information
Created a middleware function to decode the JWT token and extract the user information. The middleware function takes the JWT token from the authorization header, verifies it using the JWT secret key, and decodes the payload to get the user information. The decoded user object is added to the request object as req.user, which can be used by subsequent middleware or route handlers. This middleware will be used to authenticate and authorize requests for protected routes in the application.