Implementing useLocation Hook
This commit adds the useLocation hook to the iNotebook project. The useLocation hook is used to access the current URL location in a React component. With this hook, we are able to dynamically render components based on the current URL location. This improves the user experience and provides more flexibility in navigating through the app. The useLocation hook is now implemented in several components throughout the iNotebook project.