Add functionality to fetch notes from context
In this commit, we have added functionality to fetch notes from context in the iNotebook React project. We have used the useContext hook provided by the React Context API to access the notes data from the context. This allows us to display the notes data in different components of our application.

We have also made some modifications to the NoteState.js file to add the necessary code to fetch notes from the backend and update the state of the notes in the context.

Overall, this feature will help users to view their saved notes in the iNotebook application without any hassle.