Add global styling, mobile responsiveness, and Login component
This commit adds global styling to the application using CSS variables,
making it easier to manage and update the styles across the app. It also
introduces mobile responsiveness, ensuring that the app is usable and
looks good on mobile devices. Finally, it adds a Login component, allowing
users to sign in and access restricted content.

Fix: Wrap application with BrowserRouter to fix "useRoutes() may be used only in the context of a <Router> component" error.

In this commit, I updated the import of React Router components to use
the new `Routes` component instead of `Switch`, which is no longer
supported in v6. I also added the Login component to a route using the
`element` property.

In react-router-dom v6, "Switch" is replaced by routes "Routes".
That means you need to change the import from ➡

❌ import { BrowserRouter as Router, Switch, Route } from "react-router-dom";

to

✅ import { BrowserRouter as Router, Routes, Route } from "react-router-dom";

And then add the component inside "element" property:

<Route path="/" element={<Login/>}></Route>